Move .ui and the corresponding .py files into the rare directory under ui for clarity
