+++ rare/components/tabs/settings/about.py
@@ -3,7 +3,7 @@ import webbrowser
 from PyQt5.QtWidgets import QWidget
 
 from rare import __version__
-from rare.components.tabs.settings.about_ui import Ui_About
+from rare.ui.components.tabs.settings.about import Ui_About
 from rare.utils.utils import get_latest_version
