+++ rare/components/tabs/settings/linux.py
@@ -4,7 +4,7 @@ from PyQt5.QtWidgets import QFileDialog, QWidget
 
 from custom_legendary.core import LegendaryCore
 from rare.components.tabs.settings.dxvk import DxvkSettings, DxvkWidget
-from rare.components.tabs.settings.linux_ui import Ui_LinuxSettings
+from rare.ui.components.tabs.settings.linux import Ui_LinuxSettings
 from rare.utils.extra_widgets import PathEdit
 
 logger = getLogger("LinuxSettings")
