+++ rare/ui/components/tabs/games/game_info/game_settings.ui
