+++ rare/ui/components/tabs/settings/about.ui
