+++ rare/ui/components/tabs/settings/dxvk.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 
-# Form implementation generated from reading ui file 'resources/ui/components/tabs/settings/dxvk.ui'
+# Form implementation generated from reading ui file 'dxvk.ui'
 #
 # Created by: PyQt5 UI code generator 5.15.4
 #
