+++ rare/ui/components/tabs/settings/linux.ui
