+++ rare/ui/components/tabs/settings/rare.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 
-# Form implementation generated from reading ui file 'resources/ui/components/tabs/settings/rare.ui'
+# Form implementation generated from reading ui file 'rare.ui'
 #
 # Created by: PyQt5 UI code generator 5.15.4
 #
