+++ rare/ui/components/tabs/settings/rare.ui
